+++ data/examples/declaration/value/function/arrow/proc-applications.hs
@@ -2,6 +2,8 @@
 
 foo x = proc a -> a -< x
 
+foolr x = proc a -> x >- a
+
 bar f x =
     proc (
         y,
@@ -15,3 +17,5 @@ bar f x =
     )
 
 baz x = proc a -> a -<< x
+
+bazlr x = proc a -> x >>- a