+++ CHANGELOG.md
@@ -61,6 +61,9 @@
 * Minus signs in literal patterns are now preserved in all cases. [Issue
   733](https://github.com/tweag/ormolu/issues/733).
 
+* Added support for left-to-right arrow application. [Issue
+  737](https://github.com/tweag/ormolu/issues/737).
+
 ## Ormolu 0.1.4.1
 
 * Added command line option `--color` to control how diffs are printed.